exchanger/requirements.txt
aiogram>=2.20
python-dotenv>=0.19.0
aiohttp>=3.8.1
Babel>=2.9.1
